+++ src/fibonacci.h
@@ -1,5 +1,5 @@
 /* Very simple C example */
 /* Find the nth fibonacci number using recursion */
 
-int fibonacci(int n);
-int fast_fibonacci(int n);
+unsigned long long int fibonacci(int n);
+unsigned long long int fast_fibonacci(int n);
